refactor: clean up cursor styles and introduce TableConfig type for data table
- Removed redundant cursor pointer styles from index.css to streamline CSS.
- Added a new TableConfig type definition to enhance type safety and configuration options for the data table.
- Updated data-export and toolbar components to utilize the new TableConfig type, improving code organization and maintainability.
